continued work on #5 cleaned up PluginCommand
