+++ mcspring-api/mcspring-test/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>mcspring-api</artifactId>
         <groupId>in.kyle.mcspring</groupId>
-        <version>0.0.8</version>
+        <version>0.0.9</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
