+++ mcspring-build/mcspring-plugin-manager/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <artifactId>mcspring-starter</artifactId>
         <groupId>in.kyle.mcspring</groupId>
-        <version>0.0.8</version>
+        <version>0.0.9</version>
         <relativePath>../mcspring-starter</relativePath>
     </parent>
