+++ mcspring-examples/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>mcspring-starter</artifactId>
         <groupId>in.kyle.mcspring</groupId>
-        <version>0.0.8</version>
+        <version>0.0.9</version>
         <relativePath>../mcspring-build/mcspring-starter</relativePath>
     </parent>
