+++ mcspring-examples/simple-factions/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <artifactId>mcspring-examples</artifactId>
         <groupId>in.kyle.mcspring</groupId>
-        <version>0.0.8</version>
+        <version>0.0.9</version>
     </parent>
 
     <artifactId>simple-factions</artifactId>
