+++ pom.xml
@@ -12,7 +12,7 @@
     <groupId>in.kyle.mcspring</groupId>
     <artifactId>mcspring-parent</artifactId>
     <packaging>pom</packaging>
-    <version>0.0.8</version>
+    <version>0.0.9</version>
     <name>mcspring</name>
     <description>Adds Spring support to Bukkit plugins</description>
     <url>https://github.com/kylepls/mc-spring</url>
